Fixed CSR_RegFile_MSU.bsv so rv{32,64}mi-p-breakpoint test passes again (detail below).
Detail: The 'breakpoint' test checks the 'tdata1' CSR to see if
'Trigger' behavior has been implemented, and skips the test if not
(trivial 'pass').  Earlier, this ISA test merely read 'tdata1', and we
were returning the reset value of '0' for the 'type' field (upper 4
MSBs) indicating 'no trigger', and getting the trivial 'pass'.  The
ISA test has since then been updated to write '2' in the 'type' field
of 'tdata1' and read it back; we were returning 2, even though we
don't implement that trigger behavior, and so the test began to fail.

Fixed: always return type 0 in 'tdata1', restoring the trivial pass.
(Until we implement Trigger Module functionality, if we ever do.)