Fixed bug in CSR_Regfile_MSU.bsv (detail follows), plus small syncs with Flute.
Many CSRs have WARL/WLRL semantics: the value attempted to be written
may be transformed into an actual value written.  One example is
hard-wired CSRs where the attempted write-value is ignored.  In the
code, the fav_csr_write function should return the actual value
written; it was not doing so in all cases.  Fixed.