Updated MMU_Cache.bsv from bugfix in Flute commit dc239a30 (detail below).
Flute commit 4cd4444e has a bugfix in MMU_Cache.bsv where the
State_and_CTag_CSet was being updated to CLEAN state before observing
the first response to the burst-read.  If the response is an error
(e.g., read to an unmapped location), the state should not update to
CLEAN.  Flute commit dc239a30 has some extra tweaks on that.